Python_backend/requirements.txt
newspaper3k==0.2.8
pymysql==0.9.3
langdetect==1.0.8